exercises 4.15-4.23 done

+++ part4/bloglist/controllers/blogs.js
@@ -1,30 +1,54 @@
 const blogsRouter = require('express').Router()
-const { findByIdAndUpdate } = require('../models/blog')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
 
     response.json(blogs)
 
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', userExtractor, async (request, response) => {
+    const body = request.body
+    const user = request.user
+
+    console.log(user._id)
+
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        user: user._id,
+        likes: body.likes
+    })
 
     if ((!blog.title) && (!blog.url)) {
         response.status(400).json({ error: 'title and url missing' })
     }
     else {
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.status(201).json(savedBlog)
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    console.log(user.id)
+    console.log(blog.user.toString())
+
+    if (user.id.toString() === blog.user.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'invalid user' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
